Abstract out database connection and model bootstrapping
